requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.3.0
decorator==5.0.7
dm-tree==0.1.6
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.0
h5py==2.10.0
idna==2.10
joblib==1.0.1
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.11.2
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-probability==0.10.1
termcolor==1.1.0
threadpoolctl==2.1.0
typeguard==2.12.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
